Fix conversion warnings

+++ include/spymarine/crc.hpp
@@ -16,7 +16,7 @@ constexpr uint16_t crc(const std::span<const uint8_t> bytes) noexcept {
     crc ^= byte << 8;
     for (int i = 0; i < 8; i++) {
       if (crc & 0x8000) {
-        crc = (crc << 1) ^ poly;
+        crc = uint16_t(crc << 1) ^ poly;
       } else {
         crc <<= 1;
       }
